+++ demo/src/main/java/com/project/demo/logic/EmailService.java
@@ -4,9 +4,12 @@ import org.springframework.stereotype.Service;
 import org.springframework.mail.SimpleMailMessage;
 import org.springframework.mail.javamail.JavaMailSender;
 
+import java.util.List;
+import java.util.Optional;
+
 
 @Service
-public class EmailService {
+public class EmailService implements IService<String, String> {
 
     @Autowired
     private JavaMailSender mailSender;
@@ -17,5 +20,30 @@ public class EmailService {
         message.setSubject(subject);
         message.setText(text);
         mailSender.send(message);
+    }
+
+    @Override
+    public String save(String entity) {
+        return "";
+    }
+
+    @Override
+    public List<String> findAll() {
+        return List.of();
+    }
+
+    @Override
+    public Optional<String> findById(String s) {
+        return Optional.empty();
+    }
+
+    @Override
+    public String update(String entity) {
+        return "";
+    }
+
+    @Override
+    public void deleteById(String s) {
+
     }
 }
